% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_message.R
\name{add_message}
\alias{add_message}
\title{Add messages to a chat object.}
\usage{
add_message(chat_obj, message, role = "user")
}
\arguments{
\item{chat_obj}{A chat object created from `create_chat()`}

\item{message}{A character vector with one element. The message to add to the chat.}

\item{role}{A character vector with one element. The role of the message. Typically 'user' or 'system'.}
}
\value{
A chat object with the messages added
}
\description{
Add messages to a chat object.
}
\examples{
\dontrun{dotenv::load_dot_env()
chat_openai <- create_chat('openai', Sys.getenv('OAI_DEV_KEY'))|>
  add_model('gpt-3.5-turbo') |>
  add_params('temperature' = 0.5, 'max_tokens' = 100) |>
  add_message(
    role = 'system',
    message = 'You are a chatbot that completes texts.
    You do not return the full text.
    Just what you think completes the text.'
  ) |>
  add_message('2 + 2 is 4, minus 1 that\'s 3, ')

chat_mistral <- create_chat('mistral', Sys.getenv('MISTRAL_DEV_KEY')) |>
  add_model('mistral-large-latest') |>
  add_params('temperature' = 0.5, 'max_tokens' = 100) |>
  add_message(
    role = 'system',
    message = 'You are a chatbot that completes texts.
    You do not return the full text.
    Just what you think completes the text.'
  ) |>
  add_message('2 + 2 is 4, minus 1 that\'s 3, ')
 }
}
