% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_message.R
\name{append_message}
\alias{append_message}
\alias{append_message.tidychat}
\alias{append_message.ollama}
\alias{append_message.antropic}
\title{Append a message to a `tidychat` object.}
\usage{
append_message(chat, response, ...)

\method{append_message}{tidychat}(chat, response, ...)

\method{append_message}{ollama}(chat, response, ...)

\method{append_message}{antropic}(chat, response, ...)
}
\arguments{
\item{chat}{An object of class `tidychat`.}

\item{response}{The response as returned by `perform_query`.}

\item{...}{Ignored for future compatibility.}
}
\description{
Append a message to a `tidychat` object.
}
\section{Methods (by class)}{
\itemize{
\item \code{append_message(tidychat)}: Appends a message to a `tidychat` object.

\item \code{append_message(ollama)}: Appends a message to a `ollama` object.

\item \code{append_message(antropic)}: Appends a message to a `anthropic` object.

}}
