% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_chat.R
\name{extract_chat}
\alias{extract_chat}
\alias{extract_chat.tidychat}
\title{Prints all messages to the console and saves them invisibly}
\usage{
extract_chat(chat, silent = FALSE)

\method{extract_chat}{tidychat}(chat, silent = FALSE)
}
\arguments{
\item{chat}{A chat object created from `create_chat()`}

\item{silent}{A logical vector with one element. If TRUE, the messages are not printed to the console.}
}
\value{
A chat object with the responses added
}
\description{
Prints all messages to the console and saves them invisibly
}
\section{Methods (by class)}{
\itemize{
\item \code{extract_chat(tidychat)}: Extracts the chat messages from a `tidychat` object.

}}
\examples{
\dontrun{dotenv::load_dot_env()
chat_openai <- create_chat('openai', Sys.getenv('OAI_DEV_KEY'))|>
  add_model('gpt-3.5-turbo') |>
  add_params('temperature' = 0.5, 'max_tokens' = 100) |>
  add_message(
    role = 'system',
    message = 'You are a chatbot that completes texts.
    You do not return the full text.
    Just what you think completes the text.'
  ) |>
  add_message('2 + 2 is 4, minus 1 that\'s 3, ')
  chat_openai <- chat_openai |>
    perform_chat()
  msgs_openai <- chat_openai |> extract_chat()

chat_mistral <- create_chat('mistral', Sys.getenv('MISTRAL_DEV_KEY')) |>
  add_model('mistral-large-latest') |>
  add_params('temperature' = 0.5, 'max_tokens' = 100) |>
  add_message(
    role = 'system',
    message = 'You are a chatbot that completes texts.
    You do not return the full text.
    Just what you think completes the text.'
  ) |>
  add_message('2 + 2 is 4, minus 1 that\'s 3, ')
  chat_mistral <- chat_mistral |>
    perform_chat()
  msgs_mistral <- chat_mistral |> extract_chat()
  }
}
