% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_engine}
\alias{get_engine}
\title{Get engine from a `tidychat` object}
\usage{
get_engine(chat)
}
\arguments{
\item{chat}{An object of class `tidychat`.}
}
\description{
Get engine from a `tidychat` object
}
