% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_message.R
\name{get_messages}
\alias{get_messages}
\alias{get_messages.tidychat}
\title{Get messages from a chat object.}
\usage{
get_messages(chat)

\method{get_messages}{tidychat}(chat)
}
\arguments{
\item{chat}{An object of class `tidychat`.}
}
\description{
Get messages from a chat object.
}
\section{Methods (by class)}{
\itemize{
\item \code{get_messages(tidychat)}: Get messages from a `tidychat` object.

}}
