% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_model.R
\name{get_model}
\alias{get_model}
\alias{get_model.tidychat}
\title{Get model from a chat object.}
\usage{
get_model(chat)

\method{get_model}{tidychat}(chat)
}
\arguments{
\item{chat}{An object of class `tidychat`.}
}
\description{
Get model from a chat object.
}
\section{Methods (by class)}{
\itemize{
\item \code{get_model(tidychat)}: Gets a model from a `tidychat` object.

}}
