% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_params.R, R/utils.R
\name{get_params}
\alias{get_params}
\alias{get_params.tidychat}
\title{Get parameters from a chat object.}
\usage{
get_params(chat)

\method{get_params}{tidychat}(chat)

get_params(chat)
}
\arguments{
\item{chat}{An object of class `tidychat`.}
}
\description{
Get parameters from a chat object.

Get params from a chat object
}
\section{Methods (by class)}{
\itemize{
\item \code{get_params(tidychat)}: Add parameters to a `tidychat` object.

}}
