% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{inc_uses}
\alias{inc_uses}
\title{Increment uses from a chat object by 1}
\usage{
inc_uses(chat)
}
\arguments{
\item{chat}{An object of class `tidychat`.}
}
\description{
Increment uses from a chat object by 1
}
