% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perform_chat.R
\name{perform_query}
\alias{perform_query}
\alias{perform_query.tidychat}
\title{Perform a tidychat engine query}
\usage{
perform_query(chat, prepared_query, ...)

\method{perform_query}{tidychat}(chat, prepared_query, ...)
}
\arguments{
\item{chat}{An object of class `tidychat`.}

\item{prepared_query}{A prepared query as returned by `prepare_engine`.}

\item{...}{Ignored for future compatibility.}
}
\description{
Perform a tidychat engine query
}
\section{Methods (by class)}{
\itemize{
\item \code{perform_query(tidychat)}: Perform a tidychat engine query.

}}
