% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perform_chat.R
\name{prepare_engine}
\alias{prepare_engine}
\alias{prepare_engine.tidychat}
\alias{prepare_engine.anthropic}
\title{Prepare tidychat engine query}
\usage{
prepare_engine(chat, ...)

\method{prepare_engine}{tidychat}(chat, ...)

\method{prepare_engine}{anthropic}(chat, ...)
}
\arguments{
\item{chat}{An object of class `tidychat`.}

\item{...}{Ignored for future compatibility.}
}
\description{
Prepare tidychat engine query
}
\section{Methods (by class)}{
\itemize{
\item \code{prepare_engine(tidychat)}: Prepare a tidychat engine query.

\item \code{prepare_engine(anthropic)}: Prepare an anthropic engine query.

}}
