% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{set_uses}
\alias{set_uses}
\title{Set uses from a chat object}
\usage{
set_uses(chat, value)
}
\arguments{
\item{chat}{An object of class `tidychat`.}

\item{value}{An integer to set the uses to.}
}
\description{
Set uses from a chat object
}
