% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_params.R
\name{add_params}
\alias{add_params}
\alias{add_params.tidychat}
\title{Add a model to a chat object.}
\usage{
add_params(chat, ...)

\method{add_params}{tidychat}(chat, ...)
}
\arguments{
\item{...}{A named list of parameters to add to the chat object. Must be valid parameters from the API documentation.}

\item{chat_obj}{A chat object created from `create_chat()`}
}
\value{
A chat object with the parameters added
}
\description{
Add a model to a chat object.
}
\section{Methods (by class)}{
\itemize{
\item \code{add_params(tidychat)}: Add parameters to a `tidychat` object.

}}
\examples{
\dontrun{dotenv::load_dot_env()
chat_openai <- create_chat('openai', Sys.getenv('OAI_DEV_KEY'))|>
  add_model('gpt-3.5-turbo') |>
  add_params('temperature' = 0.5, 'max_tokens' = 100)

chat_mistral <- create_chat('mistral', Sys.getenv('MISTRAL_DEV_KEY')) |>
  add_model('mistral-large-latest') |>
  add_params('temperature' = 0.5, 'max_tokens' = 100)
}
}
